Support for specific parameter as "raw" HTTP body
In some cases, we want to perform an inference against
an image directly. We'd like to avoid needing to wrap
the image in a larger structure.

To specify which "input" the body corresponds to, set the
x-tesserai-input-body header to the name of an input.

TFI will decode the body according to its mimetype
and use its decoded value as the specified input.

If there are additional inputs you'd like to specify,
set the x-tesserai-input-json header to a JSON-encoded
object. Its fields will be combined with field defined
by the body.